APIMF-2944: initial refactor - move creation enum validation candidates from Validator to PayloadsCollector
