+++ amf-client/shared/src/test/resources/validations/reports/examples/enum-boolean-invalid.report.jvm
@@ -1,7 +1,7 @@
 Model: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml
 Profile: RAML 1.0
 Conforms? false
-Number of results: 2
+Number of results: 3
 
 Level: Violation
 
@@ -19,4 +19,12 @@ Level: Violation
   Target: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml#/web-api/end-points/%2Fpsn/head/204/parameter/header/X-PSN-Exists/scalar/schema/enum/scalar_2
   Property: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml#/web-api/end-points/%2Fpsn/head/204/parameter/header/X-PSN-Exists/scalar/schema/enum/scalar_2
   Position: Some(LexicalInformation([(10,25)-(10,30)]))
+  Location: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml
+
+- Source: http://a.ml/vocabularies/amf/validation#example-validation-error
+  Message: true is not a valid enum value
+  Level: Violation
+  Target: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml#/web-api/end-points/%2Fpsn/head/204/parameter/header/X-PSN-Exists/scalar/schema/example/default-example
+  Property: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml#/web-api/end-points/%2Fpsn/head/204/parameter/header/X-PSN-Exists/scalar/schema/example/default-example
+  Position: Some(LexicalInformation([(11,21)-(11,27)]))
   Location: file://amf-client/shared/src/test/resources/validations/enums/enum-boolean-invalid.raml
